added final commits
