+++ recipeGenerator/src/App.css
@@ -40,11 +40,13 @@
   color: #fff;
   border-radius: 0.6rem;
   border: none;
+  white-space: nowrap;
 }
 .ingredients {
   margin-left: 150px;
 }
-.ingredient h1 {
+.ingredient h1,
+h1 {
   font-weight: 600;
   line-height: 38px;
   text-indent: 30px;
@@ -152,10 +154,4 @@
   font-weight: 700;
   color: #475467;
   font-size: 18px;
-}
-#ing {
-  display: none;
-}
-#ingg {
-  display: none;
 }
